# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Swagger documentation",
        "version": "1.0"
    },
    "paths": {
...
        "/api/domains/add": {
            "post": {
                "tags": [
                    "Domains"
                ],
                "summary": "Adds a new domains from array",
                "operationId": "c057464b069c41cc29b854a1d143be1b",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "check_exist": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "domains": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "example": [
                                            "google.com",
                                            "yandex.ru",
                                            "site1.net"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Domain"
                                    }
                                },
                                "example": {
                                    "success": true,
                                    "data": {
                                        "0": {
                                            "domain": "google.com",
                                            "did": 13,
                                            "status": 200
                                        },
                                        "status_arr": {
                                            "200": "Dpmain created",
                                            "501": "Domain not registered error",
                                            "502": "Domain too short error",
                                            "503": "Domain too long error",
                                            "504": "Domain should start from letter",
                                            "505": "Wrong domain error",
                                            "506": "Domain Incorect symbols error",
                                            "508": "Domain already exist",
                                            "509": "Domain control already exist"
                                        }
                                    },
                                    "message": "Request processed"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/domains/{did}": {
            "get": {
                "tags": [
                    "Domains"
                ],
                "summary": "Get domain by did",
                "operationId": "a7f58ffa3de54611f08c707f927796b3",
                "parameters": [
                    {
                        "name": "did",
                        "in": "path",
                        "description": "did to fetch",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "domain response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Domain"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/domains/check": {
            "post": {
                "tags": [
                    "Domains"
                ],
                "summary": "Check domains from array",
                "operationId": "5e1236c0ec8d0b9c49095a5fc201eefe",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "check_exist": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "domains": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "example": [
                                            "google.com",
                                            "yandex.ru",
                                            "site1.net"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "domain response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Domain"
                                    }
                                },
                                "example": {
                                    "success": true,
                                    "data": [
                                        {
                                            "domain": "google.com",
                                            "did": 13,
                                            "error": "Domain already exist."
                                        },
                                        {
                                            "domain": "yandex.ru",
                                            "did": 18,
                                            "error": "Domain already exist."
                                        },
                                        {
                                            "domain": "site1.net",
                                            "did": 37,
                                            "error": "Domain already exist."
                                        }
                                    ],
                                    "message": "Request processed"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/domains/get_did": {
            "post": {
                "tags": [
                    "Domains"
                ],
                "summary": "Adds a new domains from array",
                "operationId": "5b087d2e43ab11897627d40db5e23014",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "domains": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "example": [
                                            "google.com",
                                            "yandex.ru",
                                            "site1.net"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "domain response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Domain"
                                    }
                                },
                                "example": {
                                    "success": true,
                                    "data": [
                                        {
                                            "domain": "ya.ru",
                                            "did": 19
                                        },
                                        {
                                            "domain": "11111",
                                            "did": false
                                        }
                                    ],
                                    "message": "Request processed"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/domains/get_domain": {
            "post": {
                "tags": [
                    "Domains"
                ],
                "summary": "Get domain domains by did array",
                "operationId": "2e91d5de6144ac3c5b9e1c1a47859232",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "did": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "example": [
                                            11,
                                            12,
                                            13
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "domain response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Domain"
                                    }
                                },
                                "example": {
                                    "success": true,
                                    "data": [
                                        {
                                            "domain": "ya.ru",
                                            "did": 19
                                        },
                                        {
                                            "domain": false,
                                            "did": 12
                                        }
                                    ],
                                    "message": "Request processed"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Domain": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Domain"
                    },
                    {
                        "required": [
                            "domain"
                        ],
                        "properties": {
                            "did": {
                                "type": "integer",
                                "format": "int64"
                            },
                            "domain": {
                                "type": "string"
                            }
                        },
                        "type": "object",
                        "example": {
                            "did": 10,
                            "domain": "google.com"
                        }
                    }
                ]
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer {token})",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}